requirements.txt
asgiref==2.3.2
astroid==1.5.3
async-timeout==3.0.0
attrs==18.1.0
autobahn==18.7.1
Automat==0.7.0
autopep8==1.3.3
certifi==2017.11.5
channels==2.1.2
chardet==3.0.4
colorama==0.3.9
confusable-homoglyphs==2.0.2
constantly==15.1.0
daphne==2.2.1
defusedxml==0.5.0
Django==2.0
django-allauth==0.34.0
django-extensions==2.0.7
django-filter==1.1.0
django-registration==2.3
django-webpack-loader==0.6.0
django-widget-tweaks==1.4.1
djangorestframework==3.7.7
djangorestframework-jwt==1.11.0
hyperlink==18.0.0
idna==2.6
incremental==17.5.0
isort==4.2.15
lazy-object-proxy==1.3.1
Markdown==2.6.11
mccabe==0.6.1
oauthlib==2.0.6
Pillow==5.1.0
psycopg2==2.7.4
pycodestyle==2.3.1
PyHamcrest==1.9.0
PyJWT==1.6.4
pylint==1.7.4
python-openid==2.2.5
python3-openid==3.1.0
pytz==2017.3
requests==2.18.4
requests-oauthlib==0.8.0
six==1.11.0
South==1.0.2
Twisted==18.7.0
txaio==18.7.1
urllib3==1.22
virtualenv==15.1.0
wrapt==1.10.11
wxPython==4.0.0b2
zope.interface==4.5.0
